object_detection_app/object_detection/requirements.txt
tensorflow
keras
opencv-python
cython
contextlib2
pillow
lxml
jupyter
numpy
matplotlib
tensorflow-object-detection-api